requirements.txt
azure-cognitiveservices-speech==1.36.0
Flask==3.0.2
gunicorn==21.2.0
pandas==2.1.2
tqdm==4.66.1